+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Remove special chars from tokens
+* Aquos contact feedback with name and booking date
 
 ### 1.5.0: June 15th, 2021
 * Fix filter bookings query